Move doJSONP to graphie-utils.ts
